Fixed linker error on windows due to a struct declared as a class.
Fixed prepackaging on Windows.

+++ cmake/prepackage.cmake
@@ -6,8 +6,11 @@
 # Remove .a files from packaging lib/ directory
 #
 
-file( GLOB STATIC_LIBS "${CMAKE_INSTALL_PREFIX}/lib/*.a" )
-file( REMOVE ${STATIC_LIBS} )
+file( GLOB STATIC_LIBS "${CMAKE_INSTALL_PREFIX}/lib/*.a" "${CMAKE_INSTALL_PREFIX}/lib/*.lib" )
+
+if ( NOT "${STATIC_LIBS}" STREQUAL "" )
+    file( REMOVE ${STATIC_LIBS} )
+endif()
 
 #
 # Remove include files from packaging directory
